Ajout token CSRF dans admin3/category

+++ src/Controller/Admin3/CategoryController.php
@@ -8,7 +8,9 @@ use App\Repository\CategoryRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Csrf\CsrfToken;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 #[Route('/admin3/category', name: 'admin3_category_')]
@@ -68,12 +70,17 @@ class CategoryController extends AbstractController
     }
 
     #[Route('/delete/{id}', name: 'delete')]
-    public function delete(Category $category, ManagerRegistry $doctrine): Response
+    public function delete(Category $category, ManagerRegistry $doctrine, Request $request, CsrfTokenManagerInterface $csrfTokenManager): Response
     {
-        $em = $doctrine->getManager();
-        $em->remove($category);
-        $em->flush();
-        $this->addFlash('success', 'Catégorie supprimée !');
+        $token = new CsrfToken('delete', $request->query->get('_csrf_token'));
+        if ($csrfTokenManager->isTokenValid($token)) {
+            $em = $doctrine->getManager();
+            $em->remove($category);
+            $em->flush();
+            $this->addFlash('success', 'Catégorie supprimée !');
+        } else {
+            $this->addFlash('danger', 'Token absent ou invalide !');
+        }
         return $this->redirectToRoute('admin3_category_index');
     }
 }
